dom/serviceworkers/test/test_post_message_source.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1142015 - Test service worker post message source </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">
  var magic_value = "MAGIC_VALUE_RANDOM";
  var registration;
  function start() {
    return navigator.serviceWorker.register("source_message_posting_worker.js",
                                            { scope: "./nonexistent_scope/" })
      .then((swr) => registration = swr);
  }

  function unregister() {
    return registration.unregister().then(function(result) {
      ok(result, "Unregister should return true.");
    }, function(e) {
      dump("Unregistering the SW failed with " + e + "\n");
    });
  }


  function testPostMessage(swr) {
    var p = new Promise(function(res, rej) {
      navigator.serviceWorker.onmessage = function(e) {
        ok(e.data === magic_value, "Worker posted the correct value.");
        res();
      }
    });

    ok(swr.installing, "Installing worker exists.");
    swr.installing.postMessage(magic_value);
    return p;
  }


  function runTest() {
    start()
      .then(testPostMessage)
      .then(unregister)
      .catch(function(e) {
        ok(false, "Some test failed with error " + e);
      }).then(SimpleTest.finish);
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.exemptFromPerDomainMax", true],
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
  ]}, runTest);
</script>
</pre>
</body>
</html>
